[refactor] move bookmark sources to pkg/browser/bookmarks
Parsers are libraries, not cmd. firefox, chromium, and ladybird each
get a package under pkg/browser so later browser import code can live
alongside them.
